Vue.js周りをリファクタリング
